Completely untested post.lisp change: use <main>
This should help not-particularly-intelligent read-it-later parsers figure out what the important part of a post page is.

As a disclaimer, I'd like to publicly note that this change is completely untested and was written by someone whose last Lisp exposure was maybe two weeks of basic Clojure five years ago. If any CSS depends on this element being a div, then this change will break things. That said, I didn't see "div.post" anywhere in the dark-mode CSS.